Architect/requirements.txt
groq
python-dotenv

# Ниже находится демонстрационное ТЗ для модуля AI Test Architect.
# Формат комментариев выбран специально, чтобы не ломать pip install -r requirements.txt.
#
# ТЗ: Интеграционная синхронизация состава изделия между T-Flex и 1С.
# Контекст:
# - T-Flex выступает мастер-системой по конструкторской структуре изделия (BOM).
# - 1С выступает мастер-системой по складу, доступности материалов и нормативам закупки.
# - Синхронизация запускается по событию "Утверждена ревизия изделия" в T-Flex.
#
# Предусловия:
# 1) В T-Flex существует изделие "Редуктор-500", ревизия R12, статус "Утверждено".
# 2) У изделия есть 3 уровня вложенности BOM:
#    - Уровень 1: Корпус, Вал, Подшипниковый узел.
#    - Уровень 2: Для Подшипникового узла -> Подшипник 6208, Стопорное кольцо.
#    - Уровень 3: Для Вала -> Шпонка, Шайба, Болт M8.
# 3) Для части позиций в 1С уже существуют карточки номенклатуры, часть отсутствует.
# 4) Включено логирование обмена и трассировка API.
#
# Бизнес-правила синхронизации:
# - Каждая позиция BOM должна маппиться в номенклатуру 1С по ключу (Артикул + Ревизия).
# - При отсутствии карточки в 1С создается новая, статус "Черновик".
# - Изменение количества в BOM должно обновлять спецификацию в 1С.
# - Удаленные из BOM позиции должны помечаться в 1С как "Неактуально", а не удаляться физически.
# - Для материалов с признаком "Критический" обновление допускается только при наличии согласования.
#
# Тестовые сценарии:
# A) Позитивный: Полная синхронизация новой ревизии R12 без ошибок.
# B) Частичный: 1 позиция без обязательного атрибута Hardness_HRC -> ожидаем ошибку валидации.
# C) Конфликт: В 1С запись заблокирована другим пользователем -> retry и корректный отчет.
# D) Производительность: BOM из 10 000 позиций должен синхронизироваться < 15 минут.
# E) Идемпотентность: Повторный запуск синхронизации не должен дублировать позиции.
#
# Ожидаемые результаты:
# - В 1С сформирована корректная спецификация по ревизии R12.
# - Все ошибки классифицированы по severity (WARN/ERROR/CRITICAL).
# - Для инцидентов сформированы рекомендации и impact на бизнес-процесс.
# - Логи содержат correlation_id для трассировки каждой позиции BOM.
